CUDA perf and correctness fixes (#333)
* Initial changes from on field perf testing.

* Cleans up CUDA compile flag handling.

* Adds missing update step for cached_field_dims_

* Fixes layer tests.

---------

+++ ateam_spatial/include/ateam_spatial/layers/in_field.hpp
@@ -28,7 +28,7 @@ namespace ateam_spatial::layers
 {
 
 CUDA_HOSTDEV float InField(
-  const int x, const int y, const FieldDimensions & field_dims,
+  const float x, const float y, const FieldDimensions & field_dims,
   const SpatialSettings & settings);
 
 }  // namespace ateam_spatial::layers
